[JIRA PLUGIN]: Get members of an Issue Security Scheme (#13)
Implement a REST api to get all members of a given Issue Security Scheme. Input is the schemeId and output is the list of all the members across all the levels in the scheme.

Note: There are subtle differences in the response compared to Jira Cloud. For example, Cloud uses projectRole as the permission name while Server/DC uses projectrole.